Clarify warning message and fix hints
W002 is raised if you have set the HijackSMTPBackend, this is ok
on staging/test, so it can be silenced there. It is bad on production.

W003 is raised if you are not using HijackSMTPBackend, this is ok
on production, so it can be silenced there. It is bad on staging/test.